+++ lib/msf/core/exploit.rb
@@ -1,7 +1,6 @@
 # -*- coding: binary -*-
 require 'msf/core'
 require 'msf/core/module'
-require 'msf/core/post'
 
 module Msf
 
@@ -30,6 +29,8 @@ end
 ###
 class Exploit < Msf::Module
 
+	require 'msf/core/post'
+	require 'msf/core/exploit/local'
 
 	##
 	# Exceptions
@@ -217,25 +218,6 @@ class Exploit < Msf::Module
 		# before being able to trigger the vulnerability.
 		#
 		Passive    = "passive"
-	end
-
-	###
-	#
-	# The local exploit class is a specialization of the exploit module class that
-	# is geared toward exploits that are performed locally.  Locally, in this
-	# case, is defined as an exploit that is realized by means other than network
-	# communication.
-	#
-	###
-	class Local < Exploit
-		include PostMixin
-
-		#
-		# Returns the fact that this exploit is a local exploit.
-		#
-		def exploit_type
-			Exploit::Type::Local
-		end
 	end
 
 	###
