+++ lib/msf/core/exploit/local.rb
@@ -1,9 +1,28 @@
+# -*- coding: binary -*-
 
-require 'msf/core/exploit/local/unix'
-require 'msf/core/exploit/local/linux_kernel'
+require 'msf/core/post_mixin'
 
 module Msf
-module Exploit::Local
+class Exploit
+
+###
+#
+# The local exploit class is a specialization of the exploit module class that
+# is geared toward exploits that are performed locally.  Locally, in this
+# case, is defined as an exploit that is realized by means other than network
+# communication.
+#
+###
+class Local < Exploit
+	include PostMixin
+
+	#
+	# Returns the fact that this exploit is a local exploit.
+	#
+	def exploit_type
+		Exploit::Type::Local
+	end
+end
 
 end
 end
